removed extra logs, removed pbs_hook_config_os_env from hook smoke and added to a new file
